refactor: Move Node subclasses to individual files
