+++ pyworkflow/pyworkflow/nodes/flow_control/__init__.py
@@ -1,1 +1,1 @@
-from .string_node import StringNode
+from .string_input import StringNode
