+++ pyworkflow/pyworkflow/nodes/io/__init__.py
@@ -1,2 +1,2 @@
-from .read_csv_node import ReadCsvNode
-from .write_csv_node import WriteCsvNode
+from .read_csv import ReadCsvNode
+from .write_csv import WriteCsvNode
